Add files via upload
Initial upload of the Computer Monitor Service. The Computer Monitor Service will break up a list of hosts available in Active Directory if the user has read access to the domain. the HostInfo, HostServiceMonitor, and StorageMonitoring classes are helper objects that give the ComputerMonitor storage, current services, and network information about the hosts. The ComputerMonitor class allows the ComputerServiceMonitor to create a ComputerMonitor object that runs on a batch of hosts. The ComputerMonitorService class will collect the results of all runs into an output file with all information gathered on each host.